Reorganized package structure and updated dependencies.
A change was made to the package structure - the Main.java file is now in the be.ylorthioir package instead of being in be.ylorthioir.dal. Additionally, dependencies for DiContainer, InitDB, and JavanaisRepository were added to Main.java in order to facilitate access to required functionalities directly from the main application.